+++ crates/clients/Cargo.toml
@@ -9,17 +9,19 @@ gadget-client-evm = { workspace = true, optional = true }
 gadget-client-tangle = { workspace = true, optional = true }
 gadget-client-networking = { workspace = true, optional = true }
 gadget-client-core = { workspace = true }
+gadget-std.workspace = true
 
 thiserror.workspace = true
 
 [features]
 default = ["std"]
 std = [
-    "gadget-client-eigenlayer?/std",
-    "gadget-client-evm?/std",
-    "gadget-client-networking?/std",
-    "gadget-client-tangle?/std",
-    "thiserror/std"
+	"gadget-client-eigenlayer?/std",
+	"gadget-client-evm?/std",
+	"gadget-client-networking?/std",
+	"gadget-client-tangle?/std",
+	"gadget-std/std",
+	"thiserror/std"
 ]
 web = ["gadget-client-tangle?/web"]
