+++ crates/clients/src/lib.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod error;
+pub use error::*;
 
 #[cfg(feature = "eigenlayer")]
 pub use gadget_client_eigenlayer as eigenlayer;
