+++ crates/macros/Cargo.toml
@@ -29,7 +29,6 @@ tokio = { workspace = true, optional = true }
 
 # Tangle re-exports
 gadget-blueprint-serde = { workspace = true, optional = true }
-gadget-client-tangle = { workspace = true, optional = true }
 tangle-subxt = { workspace = true, optional = true }
 
 # EVM re-exports
@@ -40,48 +39,48 @@ alloy-provider = { workspace = true, optional = true }
 [features]
 default = ["std"]
 std = [
-    "gadget-blueprint-proc-macro-core/std",
-    "gadget-blueprint-proc-macro/std",
-    "gadget-blueprint-serde?/std",
-    "gadget-clients?/std",
-    "gadget-context-derive/std",
-    "gadget-crypto?/std",
-    "gadget-event-listeners?/std",
-    "gadget-logging?/std",
-    "gadget-networking/std",
-    "gadget-std/std",
-    "tangle-subxt?/std",
+	"gadget-blueprint-proc-macro-core/std",
+	"gadget-blueprint-proc-macro/std",
+	"gadget-blueprint-serde?/std",
+	"gadget-clients?/std",
+	"gadget-context-derive/std",
+	"gadget-crypto?/std",
+	"gadget-event-listeners?/std",
+	"gadget-logging?/std",
+	"gadget-networking/std",
+	"gadget-std/std",
+	"tangle-subxt?/std",
 ]
 web = [
-    "gadget-clients?/web",
-    "gadget-event-listeners?/web",
-    "tangle-subxt?/web",
+	"gadget-clients?/web",
+	"gadget-event-listeners?/web",
+	"tangle-subxt?/web",
 ]
 
 tangle = [
-    "dep:gadget-blueprint-serde",
-    "dep:gadget-clients",
-    "dep:gadget-crypto",
-    "dep:gadget-event-listeners",
-    "dep:gadget-logging",
-    "dep:tangle-subxt",
-    "dep:tokio",
-    "gadget-clients?/tangle",
-    "gadget-context-derive/tangle",
-    "gadget-crypto/tangle-pair-signer",
-    "gadget-event-listeners/tangle",
+	"dep:gadget-blueprint-serde",
+	"dep:gadget-clients",
+	"dep:gadget-crypto",
+	"dep:gadget-event-listeners",
+	"dep:gadget-logging",
+	"dep:tangle-subxt",
+	"dep:tokio",
+	"gadget-clients?/tangle",
+	"gadget-context-derive/tangle",
+	"gadget-crypto/tangle-pair-signer",
+	"gadget-event-listeners/tangle",
 ]
 evm = [
-    "dep:alloy-network",
-    "dep:alloy-transport",
-    "dep:alloy-provider",
-    "dep:gadget-clients",
-    "dep:gadget-event-listeners",
-    "dep:gadget-logging",
-    "dep:tokio",
-    "gadget-clients?/evm",
-    "gadget-context-derive/evm",
-    "gadget-event-listeners/evm",
+	"dep:alloy-network",
+	"dep:alloy-transport",
+	"dep:alloy-provider",
+	"dep:gadget-clients",
+	"dep:gadget-event-listeners",
+	"dep:gadget-logging",
+	"dep:tokio",
+	"gadget-clients?/evm",
+	"gadget-context-derive/evm",
+	"gadget-event-listeners/evm",
 ]
 
 [lints]
